+++ CHANGELOG.md
@@ -17,6 +17,14 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Breaks
 
 
+## 0.6.0 - (2022-11-09)
+---
+
+### New
+* global set of ignored paths when bundling
+* remote_funciton supports optional getter function for setting boto3 Session
+
+
 ## 0.5.2 - (2022-09-08)
 ---